requirements.txt
# requirements.txt
numpy==1.26.2
chromadb==0.4.15

# other dependencies
langchain==0.3.27
langchain-community==0.3.27
langchain-groq==0.2.0
langchain-openai==0.2.5

# PDF / OCR
pdfminer.six
pdf2image
pytesseract
Pillow
lxml
tiktoken
sentence-transformers
opencv-python

# Environment
python-dotenv


